feat: Add CustomPagePagination to PostViewSet
This commit adds the `CustomPagePagination` class to the `PostViewSet` in the `post/views.py` file. The `CustomPagePagination` is a custom pagination class that sets the page size to 10, allows the page size to be specified as a query parameter, and limits the maximum page size to 50. This pagination class will be used to paginate the posts in the API response.